Switched int values to ushort Added test to print the board with given size Added board ctor with given size

+++ Logic/Engine.cs
@@ -8,8 +8,8 @@ namespace Logic
     {
         public static void TranslateCharPositionToIndices(string i_CharPos, ref int o_Col, ref int o_Row) 
         {
-            o_Col = Convert.ToInt32(i_CharPos[0]);
-            o_Row = Convert.ToInt32(i_CharPos[1]);
+            o_Col = Convert.ToUInt16(i_CharPos[0]);
+            o_Row = Convert.ToUInt16(i_CharPos[1]);
         }
     }
 }
